// src/types/stats.ts
export type Stats = {
  patk: number;
  penPct: number;
  defPct: number;

  critRatePct: number;
  critResistPct: number;

  critDamageBonusPct: number;
  critDamageReductionPct: number;

  bDmgFlat: number;
  bDmgBonusPct: number;
  rDmgFlat: number;

  sizeBonusPct: number;
  sizeReductionPct: number;

  raceBonusPct: number;
  raceReductionPct: number;

  finalPDmgBonusPct: number;
  finalPDmgReductionPct: number;

  elementCounterPct: number;
  elementEnhancePct: number;
  elementResistPct: number;

  finalDmgBonusPct: number;
  finalDmgReductionPct: number;

  pvpFinalBonusPct: number;
  pvpFinalReductionPct: number;

  pvpFlat: number;
};